refactor: use vue query instead of using store data from props
